Refactor database code.
